\name{digIt-package}
\alias{digIt-package}
\alias{digIt}
\docType{package}
\title{
\packageTitle{digIt}
}
\description{
\packageDescription{digIt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{digIt}
\packageIndices{digIt}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{digIt}

Maintainer: \packageMaintainer{digIt}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
