% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/digit-package.R
\docType{package}
\name{digit}
\alias{digit}
\alias{digit-package}
\alias{package-digit}
\title{Tools to Format Numbers for Publication}
\description{
Format numbers for publication; includes the removal of leading zeros, standardization of number of digits, and a p-value formatter.
}

