% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\name{p}
\alias{p}
\title{Format P-Values}
\usage{
p(x, alpha = getOption("digit_alpha"), digits = getOption("digit_digits"))
}
\arguments{
\item{x}{A p-value.}

\item{alpha}{The alpha cut off to use.  Defaults to .05.  Can be set
globally via: \code{options(digit_alpha = n)} where n is the alpha level.}

\item{digits}{The number of digits to use.  Defaults to 3.  Can be set
globally via: \code{options(digit_digits = n)} where n is the number of
digits beyond the decimal point to include.}
}
\value{
Returns a string of publication ready p-values.
}
\description{
Format p-values for reporting using a < or = sign if greater than alpha level.
}
\examples{
p(.05)
p(.04999999999999999)
p(.0002)
p(.0002, .001)

mod1 <- t.test(1:10, y = c(7:20))
p(mod1$p.value)

mod2 <- t.test(1:10, y = c(7:20, 200))
p(mod2$p.value)
}
\seealso{
\code{\link[digit]{f}}
}

