% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.R
\name{pc}
\alias{pc}
\title{Comma Format Large Integers}
\usage{
pc(x, big.mark = ",", ...)
}
\arguments{
\item{x}{A vector opc numbers (or string equivalents).}

\item{big.mark}{The character to include every n places.}

\item{\ldots}{Other arguments passed to \code{\link[base]{prettyNum}}.}
}
\value{
Returns a comma separted string of publication ready digits.
}
\description{
Add commas to larger integers.
}
\examples{
set.seed(4)
pc(sample(4:10, 5)^5)
## pc(sample(4:10, 5)^5, ".")
}
\seealso{
\code{\link[base]{prettyNum}}
}

